class restructure
refactor: updated the classes needed to better follow the chain of responsibility pattern as well as SOLID principles
