--- src/server/model/login/ValidateUsername.java
@@ -1,9 +1,0 @@
-package server.model.login;
-
-public class ValidateUsername implements ValidateUsers{
-    @Override
-    public boolean validate (String inputText) {
-        return true;
-    }
-    
-}
