+++ src/server/model/login/classes/LogInLogs.java
@@ -1,4 +1,4 @@
-package server.model.login;
+package server.model.login.classes;
 
 // TODO optional if time
 public class LogInLogs {
