+++ src/server/model/login/classes/ValidateUsername.java
@@ -1,0 +1,4 @@
+package server.model.login.classes;
+
+public class ValidateUsername {
+}
