+++ src/server/model/login/classes/ValidateUsers.java
@@ -1,4 +1,4 @@
-package server.model.login;
+package server.model.login.classes;
 
 public interface ValidateUsers {
     public boolean validate(String inputText);
