+++ src/server/model/login/handlers/BaseLoginHandler.java
@@ -1,0 +1,4 @@
+package server.model.login.handlers;
+
+public abstract class BaseLoginHandler {
+}
