+++ src/server/model/login/handlers/DatabaseAuthenticationHandler.java
@@ -1,0 +1,4 @@
+package server.model.login.handlers;
+
+public class DatabaseAuthenticationHandler {
+}
