+++ src/server/model/login/handlers/LoginHandler.java
@@ -1,0 +1,4 @@
+package server.model.login.handlers;
+
+public interface LoginHandler {
+}
