+++ src/server/model/session/UserSessionManager.java
@@ -1,0 +1,4 @@
+package server.model.session;
+
+public class UserSessionManager {
+}
